# buffer reads from sha3