When added atop the sha3 alloc improvements in https://go.dev/cl/544817:

```
goos: darwin
goarch: arm64
pkg: filippo.io/mlkem768
            │   4b2bb98   │           e9cd519           │
            │   sec/op    │   sec/op     vs base        │
SampleNTT-8   1.994µ ± 0%   1.434µ ± 0%  -28.08% (n=50)
KeyGen-8      48.03µ ± 0%   41.96µ ± 1%  -12.63% (n=50)
Encaps-8      66.51µ ± 0%   59.98µ ± 0%   -9.81% (n=50)
Decaps-8      79.32µ ± 0%   72.91µ ± 0%   -8.08% (n=50)
RoundTrip-8   200.0µ ± 0%   187.1µ ± 0%   -6.43% (n=50)
geomean       39.89µ        34.55µ       -13.39%
```