@josharian 2023-11-26
Another thing you could patch atop this, if you want, that shaves another ~1.5% on arm64: https://github.com/josharian/mlkem768/commit/db20f6cdd01790542d4456124cf33bf351f3be4a