@FiloSottile 2023-12-05
Awesome! I had actually tried buffering and somehow I didn't manage to get an effect out of it (2e6a3dff2e6cd860e36a8b439c1765e14c5d3a0f). Thank you ✨ 

```
goos: linux
goarch: amd64
pkg: filippo.io/mlkem768
cpu: Intel(R) Core(TM) i5-7400 CPU @ 3.00GHz
name         old time/op  new time/op  delta
SampleNTT-4  3.13µs ± 1%  2.22µs ± 0%  -29.03%  (p=0.000 n=10+9)
KeyGen-4     85.9µs ± 1%  73.9µs ± 1%  -13.95%  (p=0.000 n=10+10)
Encaps-4      123µs ± 1%   112µs ± 0%   -8.99%  (p=0.000 n=10+7)
Decaps-4      148µs ± 0%   136µs ± 1%   -7.90%  (p=0.000 n=9+10)
RoundTrip-4   360µs ± 0%   339µs ± 0%   -5.86%  (p=0.000 n=9+10)
```