@FiloSottile 2023-12-05
Huh, weird, I consistently get a strongly negative effect from https://github.com/josharian/mlkem768/commit/db20f6cdd01790542d4456124cf33bf351f3be4a and arm64 though.

```
name         old time/op  new time/op  delta
SampleNTT-8  2.57µs ± 0%  2.81µs ± 0%  +9.40%  (p=0.000 n=10+10)
KeyGen-8     71.3µs ± 2%  74.6µs ± 4%  +4.50%  (p=0.000 n=10+10)
Encaps-8      101µs ± 1%   104µs ± 1%  +3.26%  (p=0.000 n=9+8)
Decaps-8      121µs ± 1%   124µs ± 1%  +2.76%  (p=0.000 n=10+10)
RoundTrip-8   313µs ± 0%   319µs ± 0%  +1.74%  (p=0.000 n=10+10)
```

While on amd64 the effect is positive.

```
name         old time/op  new time/op  delta
SampleNTT-4  2.31µs ± 1%  2.16µs ± 1%  -6.53%  (p=0.000 n=10+10)
KeyGen-4     76.9µs ± 2%  73.8µs ± 1%  -4.12%  (p=0.000 n=10+10)
Encaps-4      113µs ± 1%   112µs ± 1%  -1.19%  (p=0.000 n=9+10)
Decaps-4      139µs ± 1%   136µs ± 1%  -2.07%  (p=0.000 n=10+10)
RoundTrip-4   342µs ± 0%   335µs ± 0%  -2.17%  (p=0.000 n=8+9)
```

I just patched the commit on top of this PR, on top of main, with a replace for https://go.dev/cl/544817.