@FiloSottile 2023-12-05
🤦 I had not committed the replace directive, so what I was comparing in https://github.com/FiloSottile/mlkem768/pull/1#issuecomment-1840547660 was fd793ab-without-CL544817 (in a clean worktree) vs fd793ab+db20f6c+CL544817.

That explains _some_ of the weirdness.

However, I guess the question now is why CL 544817 seems to have a negative effect on arm64.