@FiloSottile 2023-12-05
Indeed, CL 544817 (and the previous in the stack) have a negative performance effect on arm64, and a positive one on amd64, even after all these patches.

```
goos: darwin
goarch: arm64
pkg: filippo.io/mlkem768
name         old time/op  new time/op   delta
SampleNTT-8  2.51µs ± 0%   2.80µs ± 0%  +11.57%  (p=0.000 n=7+10)
KeyGen-8     69.3µs ± 1%   72.0µs ± 1%   +3.87%  (p=0.000 n=9+9)
Encaps-8     96.1µs ± 1%  100.2µs ± 1%   +4.25%  (p=0.000 n=9+10)
Decaps-8      116µs ± 1%    120µs ± 1%   +3.65%  (p=0.000 n=10+10)
RoundTrip-8   299µs ± 0%    310µs ± 0%   +3.97%  (p=0.000 n=10+9)
```

```
goos: linux
goarch: amd64
pkg: filippo.io/mlkem768
cpu: Intel(R) Core(TM) i5-7400 CPU @ 3.00GHz
name         old time/op  new time/op  delta
SampleNTT-4  2.26µs ± 1%  2.15µs ± 0%  -4.62%  (p=0.000 n=9+8)
KeyGen-4     72.9µs ± 1%  70.8µs ± 1%  -2.90%  (p=0.000 n=9+9)
Encaps-4      109µs ± 1%   107µs ± 1%  -1.89%  (p=0.000 n=10+10)
Decaps-4      133µs ± 2%   131µs ± 1%  -2.00%  (p=0.000 n=10+10)
RoundTrip-4   328µs ± 1%   323µs ± 1%  -1.36%  (p=0.000 n=10+10)
```

Anyway. These patches have a positive impact regardless. Submitting them, thank you!