@josharian 2023-12-05
> I consistently get a strongly negative effect from https://github.com/josharian/mlkem768/commit/db20f6cdd01790542d4456124cf33bf351f3be4a and arm64 though.

Huh. Maybe your arm64 isn't happy with unaligned reads. Shrug. Computers are hard. :) Glad some of it was useful.