@FiloSottile 2023-12-06
> > I consistently get a strongly negative effect from [josharian@db20f6c](https://github.com/josharian/mlkem768/commit/db20f6cdd01790542d4456124cf33bf351f3be4a) and arm64 though.
> 
> Huh. Maybe your arm64 isn't happy with unaligned reads. 

Nono I think I was actually measuring the (still puzzling) negative effect of CL 544817 by mistake. https://github.com/FiloSottile/mlkem768/pull/1#issuecomment-1840642882